requirements.txt
streamlit
requests
matplotlib
python-dotenv
fpdf
geocoder
